Next Market完成

+++ pages/api/item/delete/[id].js
@@ -1,15 +1,21 @@
 // pages/api/item/delete/[id].js
 
+import auth from "../../../../utils/auth";
 import connectDB from "../../../../utils/database";
 import { ItemModel } from "../../../../utils/schemaModels";
 
 const deleteItem = async (req, res) => {
     try {
         await connectDB();
-        await ItemModel.deleteOne({ _id: req.query.id })
-        return res.status(200).send({
-            message: "アイテム削除成功"
-        });
+        const singleItem = await ItemModel.findById(req.query.id);
+        if (singleItem.email === req.body.email) {
+            await ItemModel.deleteOne({ _id: req.query.id })
+            return res.status(200).send({
+                message: "アイテム削除成功"
+            });
+        }
+
+        throw new Error();
     } catch (error) {
         return res.status(400).json({
             message: "アイテム削除失敗"
@@ -17,4 +23,4 @@ const deleteItem = async (req, res) => {
     }
 };
 
-export default deleteItem;
+export default auth(deleteItem);
